Remove cropping of featured content image until we can support it proper
