+++ packages/gdl-frontend/pages/books/_book.js
@@ -12,6 +12,7 @@ import NextLink from 'next/link';
 import getConfig from 'next/config';
 import styled from 'react-emotion';
 import copyToClipboard from 'copy-to-clipboard';
+import { imageUrl } from 'gdl-image';
 import {
   Menu,
   MenuItem,
@@ -118,7 +119,9 @@ class BookPage extends React.Component<Props> {
         <Head
           description={book.description}
           title={book.title}
-          image={book.coverImage && book.coverImage.url}
+          image={
+            book.coverImage && imageUrl(book.coverImage, { aspectRatio: 0.81 })
+          }
         >
           <BookJsonLd book={book} />
         </Head>
